add permission

+++ src/main/java/com/project/logistic_management_2/service/truck/TruckServiceImpl.java
@@ -2,6 +2,8 @@ package com.project.logistic_management_2.service.truck;
 
 import com.project.logistic_management_2.dto.truck.TruckDTO;
 import com.project.logistic_management_2.entity.Truck;
+import com.project.logistic_management_2.enums.PermissionKey;
+import com.project.logistic_management_2.enums.PermissionType;
 import com.project.logistic_management_2.exception.def.NotFoundException;
 import com.project.logistic_management_2.mapper.truck.TruckMapper;
 import com.project.logistic_management_2.repository.truck.TruckRepo;
@@ -17,9 +19,10 @@ public class TruckServiceImpl extends BaseService  implements TruckService {
 
     private final TruckRepo repository;
     private final TruckMapper mapper;
-
+    private final PermissionType type = PermissionType.TRUCKS;
     @Override
     public TruckDTO createTruck(TruckDTO truckDTO) {
+        checkPermission(type, PermissionKey.WRITE); // check quyền tạo
         Truck truck = mapper.toTruck(truckDTO);
         repository.save(truck);
         return mapper.toTruckDTO(truck);
@@ -27,21 +30,24 @@ public class TruckServiceImpl extends BaseService  implements TruckService {
 
     @Override
     public List<TruckDTO> getAllTrucks() {
+        checkPermission(type, PermissionKey.VIEW); // Kiểm tra quyền xem
         List<Truck> trucks = repository.getAllTrucks();
         return trucks.stream().map(mapper::toTruckDTO).toList();
     }
 
     @Override
-    public List<TruckDTO> getTrucksByType(Integer type) {
-        List<Truck> trucks = repository.getTrucksByType(type);
+    public List<TruckDTO> getTrucksByType(Integer Type) {
+        checkPermission(type, PermissionKey.VIEW);
+        List<Truck> trucks = repository.getTrucksByType(Type);
         if (trucks.isEmpty()) {
-            throw new NotFoundException("Không tìm thấy xe với loại: " + type);
+            throw new NotFoundException("Không tìm thấy xe với loại: " + Type);
         }
         return trucks.stream().map(mapper::toTruckDTO).toList();
     }
 
     @Override
     public TruckDTO getTruckByLicensePlate(String licensePlate) {
+        checkPermission(type, PermissionKey.VIEW);
         Truck truck = repository.getTruckByLicensePlate(licensePlate)
                 .orElseThrow(() -> new NotFoundException("Không tìm thấy xe với biển số: " + licensePlate));
         return mapper.toTruckDTO(truck);
@@ -49,6 +55,7 @@ public class TruckServiceImpl extends BaseService  implements TruckService {
 
     @Override
     public TruckDTO updateTruck(Integer id, TruckDTO truckDTO) {
+        checkPermission(type, PermissionKey.WRITE);
         Truck truck = repository.getTruckById(id)
                 .orElseThrow(() -> new NotFoundException("Không tìm thấy thông tin xe cần tìm!"));
 
@@ -59,9 +66,11 @@ public class TruckServiceImpl extends BaseService  implements TruckService {
 
     @Override
     public long deleteTruck(Integer id) {
-       Truck truck = repository.getTruckById(id)
-               .orElseThrow(() -> new NotFoundException("Không tìm thấy thông tin xe"));
-       return repository.delete(id);
+        checkPermission(type, PermissionKey.DELETE); // Kiểm tra quyền xóa
+        Truck truck = repository.getTruckById(id)
+                .orElseThrow(() -> new NotFoundException("Không tìm thấy thông tin xe"));
+        return repository.delete(id);
     }
+
 }
